feat: new web socket example

+++ examples/09_ticker_req/main.go
@@ -22,11 +22,11 @@ func main() {
 		log.Fatalf("Connect failed: %v", err)
 	}
 
-	if err := client.Subscribe(context.Background(), cli, int32(constant.Market_US), "NVDA", []constant.SubType{constant.SubType_Ticker}); err != nil {
+	if err := client.Subscribe(context.Background(), cli, constant.Market_US, "NVDA", []constant.SubType{constant.SubType_Ticker}); err != nil {
 		log.Fatalf("Subscribe failed: %v", err)
 	}
 
-	tickers, err := client.GetTicker(context.Background(), cli, int32(constant.Market_US), "NVDA", 20)
+	tickers, err := client.GetTicker(context.Background(), cli, constant.Market_US, "NVDA", 20)
 	if err != nil {
 		log.Fatalf("GetTicker failed: %v", err)
 	}
